move store and reducers
