+++ tic-tac-toe/src/app/screens/LoginContainer/index.js
@@ -3,10 +3,10 @@ import React, { Component } from 'react';
 import Login from './layout';
 
 class LoginContainer extends Component {
-  submit = values => windows.alert(JSON.stringify(values, null, 4));
+  submit = values => alert(JSON.stringify(values, null, 4));
 
   render() {
-    <Login onSubmit={this.submit} />;
+    return <Login onSubmit={this.submit} />;
   }
 }
