+++ tic-tac-toe/src/index.js
@@ -3,26 +3,14 @@ import ReactDOM from 'react-dom';
 
 import './scss/index.scss';
 import { Provider } from 'react-redux';
-import { createStore, combineReducers } from 'redux';
-import { reducer as formReducer } from 'redux-form';
 
+import store from './redux/store';
 import registerServiceWorker from './registerServiceWorker';
-import LoginContainer from './app/screens/LoginContainer/layout';
-
-const reducers = {
-  form: formReducer
-};
-
-const reducer = combineReducers(reducers);
-
-const store = createStore(
-  reducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
+import Login from './app/screens/LoginContainer';
 
 ReactDOM.render(
   <Provider store={store}>
-    <LoginContainer />
+    <Login />
   </Provider>,
   document.getElementById('root')
 );
