+++ tic-tac-toe/src/redux/store.js
@@ -1,0 +1,10 @@
+import { createStore } from 'redux';
+
+import reducer from './gameReducer/reducers';
+
+const store = createStore(
+  reducer,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
+
+export default store;
